backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-docx==1.1.2
pypdf==5.1.0
pdf2image==1.17.0
Pillow==11.0.0
python-dotenv==1.0.1
